feat(aws): More Redshift resources
