--- plugins/source/aws/resources/services/redshift/clusters_fetch.go
@@ -1,52 +1,0 @@
-package redshift
-
-import (
-	"context"
-	"fmt"
-
-	"github.com/aws/aws-sdk-go-v2/aws"
-	"github.com/aws/aws-sdk-go-v2/service/redshift"
-	"github.com/aws/aws-sdk-go-v2/service/redshift/types"
-	"github.com/cloudquery/cloudquery/plugins/source/aws/client"
-	"github.com/cloudquery/plugin-sdk/schema"
-)
-
-func fetchRedshiftClusters(ctx context.Context, meta schema.ClientMeta, parent *schema.Resource, res chan<- any) error {
-	var config redshift.DescribeClustersInput
-	c := meta.(*client.Client)
-	svc := c.Services().Redshift
-	for {
-		response, err := svc.DescribeClusters(ctx, &config)
-		if err != nil {
-			return err
-		}
-		res <- response.Clusters
-		if aws.ToString(response.Marker) == "" {
-			break
-		}
-		config.Marker = response.Marker
-	}
-	return nil
-}
-
-func resolveRedshiftClusterLoggingStatus(ctx context.Context, meta schema.ClientMeta, resource *schema.Resource, c schema.Column) error {
-	r := resource.Item.(types.Cluster)
-
-	cl := meta.(*client.Client)
-	svc := cl.Services().Redshift
-	cfg := redshift.DescribeLoggingStatusInput{
-		ClusterIdentifier: r.ClusterIdentifier,
-	}
-	response, err := svc.DescribeLoggingStatus(ctx, &cfg)
-	if err != nil {
-		return err
-	}
-
-	return resource.Set(c.Name, response)
-}
-
-func resolveClusterArn() schema.ColumnResolver {
-	return client.ResolveARN(client.RedshiftService, func(resource *schema.Resource) ([]string, error) {
-		return []string{fmt.Sprintf("cluster:%s", *resource.Item.(types.Cluster).ClusterIdentifier)}, nil
-	})
-}
