+++ plugins/source/aws/resources/services/redshift/event_subscriptions.go
@@ -1,6 +1,12 @@
 package redshift
 
 import (
+	"context"
+	"fmt"
+
+	"github.com/aws/aws-sdk-go-v2/aws"
+	"github.com/aws/aws-sdk-go-v2/aws/arn"
+	"github.com/aws/aws-sdk-go-v2/service/redshift"
 	"github.com/aws/aws-sdk-go-v2/service/redshift/types"
 	"github.com/cloudquery/cloudquery/plugins/source/aws/client"
 	"github.com/cloudquery/plugin-sdk/schema"
@@ -34,4 +40,39 @@ func EventSubscriptions() *schema.Table {
 			},
 		},
 	}
+}
+
+func fetchRedshiftEventSubscriptions(ctx context.Context, meta schema.ClientMeta, parent *schema.Resource, res chan<- any) error {
+	cl := meta.(*client.Client)
+	svc := cl.Services().Redshift
+	var params redshift.DescribeEventSubscriptionsInput
+	params.MaxRecords = aws.Int32(100)
+	for {
+		result, err := svc.DescribeEventSubscriptions(ctx, &params)
+		if err != nil {
+			return err
+		}
+		res <- result.EventSubscriptionsList
+		if aws.ToString(result.Marker) == "" {
+			break
+		}
+		params.Marker = result.Marker
+	}
+	return nil
+}
+
+func resolveEventSubscriptionARN(ctx context.Context, meta schema.ClientMeta, resource *schema.Resource, c schema.Column) error {
+	cl := meta.(*client.Client)
+	sub := resource.Item.(types.EventSubscription)
+	return resource.Set(c.Name, eventSubscriptionARN(cl, *sub.CustSubscriptionId))
+}
+
+func eventSubscriptionARN(cl *client.Client, name string) string {
+	return arn.ARN{
+		Partition: cl.Partition,
+		Service:   string(client.RedshiftService),
+		Region:    cl.Region,
+		AccountID: cl.AccountID,
+		Resource:  fmt.Sprintf("eventsubscription:%s", name),
+	}.String()
 }
